backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
websockets==12.0
httpx>=0.25.2,<1.0.0  # Updated to allow newer versions that support Python 3.11
# Removed asyncio-subprocess as it's not a real package - using asyncio's built-in instead
google-cloud-storage==2.10.0
pydantic-settings==2.1.0
tenacity==8.2.3
jinja2==3.1.2
prometheus-client==0.19.0
psutil==5.9.6
PyJWT==2.8.0
